Sequence: Fix removeLocalReferencePosition inconsistency (#10801)
Description
In a recently added method to sequence, removeLocalReferencePosition, the parameter type is reference type is ReferencePostition, when it should be LocalReferencePosition. Currently the types are equivalent as they only differ by optional properties, but there are changes coming to LocalReferencePosition which will deviate from ReferencePostition. Specifically, the change will be to incorporate TrackingGroups into LocalReferencePosition (AB#151)