Update ingresses and ingress controllers
With the upgrade of the AKS cluster from k8s 1.19 to 1.23, the ingress controllers broke because they use a pretty old image that still references some removed APIs.
As part of updating that image, this change updates the Ingress resources to use the newer way of "connecting" to controllers (i.e. IngressClass). I think this could be simplified a bit if we used namespaces in the AKS cluster.