Add more generic handling of release groups in fluid build tools (#10826)
The build tools handling of release groups (called monorepos in the code) is haphazard. This PR standardizes around the MonoRepoKind enum, and plumbs that enum through the stack.

New monorepos (release groups) can be added by adding new enum values and configuring them in the fluid-build section of the root package.json.