Add new flub CLI and support for setting the "current version" across release groups (#10866)
flub is a build and release tool for the Fluid Framework GitHub repositories. flub is intended to replace the existing
fluid build-tools, primarily by reusing existing build-tools functionality and wrapping it in a more consistent,
maintainable CLI using oclif.

Build-tools contains a lot of functionality, like policy-check, bump version, etc. However, it's very inconsistent in how command line parsing is done, and the arguments between commands is not consistent. This makes reasoning about the code difficult, and it makes maintenance challenging. When someone wants to add a new command, or update existing commands, there's no infrastructure or common patterns, so people just add what's needed to make their scenario work.

Thus the goal of this change is threefold:

1. Provide a clear pattern to add new commands, re-use existing argument parsing, etc. without completely ripping out the current build-tools
1. Handle CLI input and argument parsing in a consistent way.
1. Re-use existing code from build-tools wherever possible. In most cases the lib-type code should continue to live in build-tools and be called from flub.