docs: Remove usages of `@see` tag, which is not TSDoc compliant (#10901)
Replaces usages of the `@see` tag in source-code comments. That tag is a part of the JSDoc spec, but not of TSDoc, which is what we use. Resulting documentation generation does not respect this tag, and ends up omitting referenced content.

This change replaces usages of that block with appropriate replacement syntax. 

Also includes a couple of related misc. fixes, and updates whitespace formatting for consistency.