Support yarn workspaces in build-tools (#10961)
The build tools make assumptions that release groups (monorepos) are defined in lerna.json. As we consider moving to yarn, the build-tools need to load packages from different locations. This change adds logic into the common build-tools infrastructure to load packages from lerna.json if available, but to fall back to the workspaces field in package.json if lerna.json is missing.

The version for the release group will be loaded from lerna.json if present, otherwise the release group root package.json version will be used. This allows us to continue to version packages without using the root package.json to define it. That seems like it will be helpful, but I can't think of a specific use-case.