Update nginx ingress controllers, and related changes (#10888)
With the upgrade of the AKS cluster from k8s 1.19 to 1.23, the ingress controllers broke because they use a pretty old image that still references some removed APIs.
As part of updating that image, this change updates the Ingress resources to use the newer way of "connecting" to controllers (i.e. IngressClass). I think this could be simplified a bit if we used namespaces in the AKS cluster.
This also replaces the manual deployment of the nginx ingress controller with a Helm chart.