Initial support for internal package version scheme (#11060)
Adds capability to manage versions using the Fluid internal version scheme.

The Fluid internal version scheme consists of two semver "triplets" of major/minor/patch. The first triplet is called
the public version, and is stored in the typical semver positions in the version string.

The second triplet is called the internal version, and is found at the end of the pre-release section of the
version string.