MergeTree: Organize and Deprecated Collections (#11137)
Description
The merge tree package exports a number of collection classes that it shouldn't export. A number should be internal to merge tree itself and not exported at all, some are used across packages, and some are unused in merge tree, but used in sequence.

Note: I am intentionally not including a breaking note here. These classes are not fundamental to the merge tree api surfaces, and are only exported for erroneous legacy reasons. Since they are not on the api surface, i do not expect any outside usage