Delete unused code. (#11151)
It was disabled November 2020 - see https://github.com/microsoft/FluidFramework/pull/4285/files. Even before that, it was problematic - I think we kept it around for short period of time, it was causing problems, so we disabled it.
It feels safe to remove it. I think GA was spring of 2021, and early preview was around that timeframe when we disabled it.