MergeTree: Re-Enable LocalReference Tombstones (#11142)
## Description
For upcoming changes related to undo redo we need to use the tombstone positions for references when they slide. The reason we need the tombstone positions is to maintain relative ordering between local references to removed segments. This will have no impact on other usages of local references which do not care about the internal ordering of the reference themselves.

Additionally, i'm made some simplifications to the how we slide reference in merge tree, especially around the offset. In a side scenario we will always slide to one end of a segment or the other, and this is easy to calculate using the ordinal, the code now takes advantage of this. 

There we also some issue in the tombstone handling which are un-covered by new test work related to local references. The code was updated to fix bug around those case.

## Other information or known dependencies

[AB#212](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/212)