Isolate merge tree to not export in the future (#11162)
Description
This change has no functional changes. It isolates the merge tree class and deprecates it, which will allow us to remove it from the exports in the next major release. The primary interface for merge tree is actually the client class, which is unchanged here. Overall, this will increase our ability to make changes to underlying data structure, MergeTree with needing breaking changes.