Wait to raise Connected event for Read connection until container is caught up (#9377)
The desired outcome is that regardless of connection mode (which is an implementation detail - we use read connection even if we have write permissions sometimes), the semantics of the "connected" event are consisted:  It should mean that you're connected and caught up.  This has been the case for write but not for read connections.

This change is "dark" meaning no behavior change is expected with this PR, but there's a setting "Fluid.Container.CatchUpBeforeDeclaringConnected" which can be set to true to enable the new behavior.  Once this is tested and permanently enabled, we should be able to deprecate `waitContainerToCatchUp` in favor of just responding to the `connected` event.