Move CLI tools to fluid-tools scope and invert some dependency directions (#11215)
Moves some functions from the build-tools package to the version-tools package, which is a better package for these functions. In order to do this, the dependency direction needs to be reversed, so build-tools depends on version-tools.

The version-tools package will also need to be published, so this PR moves (renames) the CLI packages to use the @fluid-tools scope.

Finally, the build-tools build pipeline is updated to use pre-release versioning in CI like the rest of our multi-package release groups do. This enables us to treat build-tools just like any other release group.