Gc config tests (#11226)
Add some unit tests for GC configuration that happens in the constructor.

This is just a start, and only covers part of the logic in the constructor - mostly around the properties that are persisted. There's lots more ground to cover when you also consider overrides that apply within the session.