Improve comment in SummaryManager.getShouldSummarizeState() (#11480)
Fix for [AB427](https://dev.azure.com/fluidframework/internal/_workitems/edit/427)
Adding 2 new Summary Reasons to getShouldSummarizeState: notElectedParent and notElectedClient
to replace the original parentShouldNotSummarize one. 

## Description
Instead of parentShouldNotSummarize, we will be using 
- **notElectedParent** - to indicate there was an election and a new client has been elected  
- **notElectedClient** - to indicate that we are not the current elected client and we are not running the (last) summary 

Concepts: 
electedClient - is the non-interactive summarizer client if one exists. If not, then electedClient is equal to
electedParent. If electedParent === electedClient, this is the signal for electedParent to spawn a new
electedClient. 

electedParent - is the interactive client that has been elected to spawn a summarizer.


## Does this introduce a breaking change?

yes as we remove the parentShouldNotSummarize reason. 

## Other information or known dependencies

I talked to Andrei and as the use of the getShouldSummarizeState is restricted to the containerRuntime, I believe there will be no  backward compatibility issue.