Move TelemetryNullLogger to telemetry-utils (#11498)
Doesn't seem to make sense for this single logger to live in common-utils -- moving this will make it easier to find and consolidate related code, and I think it aligns with the similar movement in #9016.  This change would also update all existing imports in our repo in preparation for the eventual removal.