Fix PendingStateManager replayPendingStates (#11559)
For more information about how to contribute to this repo, visit [this page](https://github.com/microsoft/FluidFramework/blob/main/CONTRIBUTING.md).

## Description

There were some cases where "flush" was not called when in `FlushMode.TurnBased`, causing some ops to not get replayed by the `PendingStateManager`. The "flush" calls were most likely not made because the ops were created when the container was not connected.