MergeTree: Make SortedSegmentSet Iterative and Add Some Tests (#11605)
Description
This is related to #11603 and #1009. Like the merge tree walk functions the recursiveness here was also causing perf issue when trying during undo, which frequently adds and removes items from a SortedSegmentSet via a tracking group. There were no directly test for SortedSegmentSet, as it was mainly tested via tracking groups, so adding a couple tests. Additionally, using a SortedSegmentSet with an object that provided a segments {segment: ISegment} was broken previously. This change also fixes that issue by allowing multiple elements with a match ordinal.

AB#151