Fix GeneralError test (#11632)
Test telemetry indicates that this test has`"R11s fetch error: {\"code\":\"NotFoundError\", ...` during snapshot fetch at loading, This test is expected to throw on connection instead. Additionally, the test doesn't fail from having a different error from expected.

The root cause is that we didn't pass in the resolved URL, because t9s/r11s/afr assign container id and don't use the "name hint".
Fixed it by pass it into `createContainerUrl`.

Additionally, change the test to be more strict about what kind of error it is expecting, so that it doesn't hide hidden issue in the future.