MergeTree: Reimplement list (#11625)
Description
The pre-existing list class was inconsistent in terms of typing, naming and semantics which made it very hard to use and extend. It also had a performance issue where whether an item was in a list, which is frequently used by local references was O(N).

This change fully re-implements the list class. It remains a circularly linked list, but separates the list and node implementation for better typing, and adds a reference to the head node to each list node to support O(1) determining if a node is in a particular list. Additionally, all methods have been implemented to match array methods in terms of shape and semantics as closely as possible to ease usage for those familiar with javascript arrays. The new implementation does regress the unused clear method, which is now O(N) rather than O(1) as each not must be detached from the head node as well.