Rework TaskManager task locking APIs (#11530)
## Description

This PR reworks `TaskManager.lockTask()` to `volunteerForTask()` and introduces `TaskManager.subscribeToTask()`.

### `volunteerForTask()`
This API largely remains unchanged from `lockTask()`. One major difference is that the return type was changed form `Promise<void>` to `Promise<boolean>`. This was done to prepare for the API eventually return `false` in certain circumstances, mainly if the task has already been completed (see internal design doc linked below).

The main motivation behind the rename was to be more similar to `subscribeToTask()` and to emphasize that there is no guarantee that this API will ever obtain the task lock.

### `subscribeToTask()`

This API is similar to `volunteerForTask()` with a few key differences. First, this API will continuously try to lock the given task. If the lock is lost due to losing connection, then we will automatically try to re-enter the queue upon reconnection. This is ideal for ongoing or asynchrounous tasks that do not have a definitive end. Second, this API does not have a return value, since calling this API once can result in the task lock being assigned and lost multiple times. To effectively use this API, the caller must rely on listening to the `assigned` and `lost` events to determine if the task lock is assigned to the client.

### `subscribed()`

This API simply returns a boolean to indicate if the client is currently subscribed to a given task.

### `haveTaskLock()` / `assignedTask()`

This PR also renames `haveTaskLock()` to `assignedTask()`. This was done to be more consistent with the `assigned` event, as well as the new APIs. There are no functional changes to this API.

### Clicker Demo
This PR also updates the clicker demo to demonstrate a case where `subscribeToTask()` is useful and how someone may implement it in their application. The previous iteration of clicker demo would always try obtaining the task lock no matter what. If the lock was lost the client would simply re-enter the queue if possible. This is a good opportunity to demonstrate `subscribeToTask()` and the coding patterns associated with it.

## Reviewer Guidance

Would appreciate feedback on:
- New API names
- New coding patterns
- Implementation details
- Return values of the above functions

## Other information or known dependencies

[MSFT Internal Design Doc](https://microsoft-my.sharepoint-df.com/:w:/p/scottnorton/ET4Tb-f-eB1Nnilpg58cBm4BMvnkkH7FN04195coc2HSEA?e=0hJkkv)

[AB#1466](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/1466)