Change logger signature. (#11636)
The logger interface is used to enable existing build-tools to use the same logging infra as the new tools with minimal changes. After using the interface for a few weeks, I think these changes to the method names remove a lot of duplicate words in code. For example, logger.logVerbose() becomes logger.verbose(), or even log.verbose() if desired.