r11s-driver: reduce API export surface (#11756)
## Description

No external consumers of R11s driver should rely on internals like DocumentService or DocumentDeltaStorage. All entry should be through DocumentServiceFactory. There are a few extra exports like DefaultTokenProvider and token types that are necessary for consumers.

Unfortunately, DocumentStorageService is shared with local-driver so I've implemented a simplified LocalDocumentStorageService class within `local-driver` that is very reminiscent of the r11s-driver DocumentStorageService from 2020.

This should help reduce the need to add a hodge-podge of optional params in internal driver classes for the sake of backwards compatibility.

## Does this introduce a breaking change?

Removes export of some internal class declarations.