Update website code to use the new `api-markdown-documenter` package for generating API docs (#11754)
## Summary

This PR updates the website (/docs) build to use `@fluid-tools/api-markdown-documenter` to generate API documentation from the `.api.json` files generated by API-Extractor.

### Formatting
This version of the package attempts to maintain parity with the current rendering as much as is reasonable. The only exceptions are changes where I felt the new behavior was superior and maintaining parity would require more work than keeping the better behavior. There are also a handful of other minor improvements that were easy enough to make, that I put them in v1. Changes include:
- Dynamic column rendering in tables for properties that may or may not appear (e.g. the `Modifiers` column - we now only render this column in tables if any entry has a value to display).
- Adds rendering for index signatures (which API-Documenter missed)
- Adds `Return Type` column to function/method tables
- Adds `Type` column to variable/parameter tables
- Sub-headings are used in more places (e.g. where we simply used bold formatting to denote sub-sections, including `Signature`)
- More explicit heading identifiers to ensure links can be uniquely resolved
- The formatting of the injected Hugo frontmatter has been prettified to make it a bit more human-readable.

> For reference, I added a couple of before and after documents under `__pr-temp` that demonstrates some of the formatting changes. You can view the raw markdown contents as well as the preview rendering here.

### Related changes

This PR also fixes links and comment formatting issues in a number of places. Some are fixes related to the new infra, but others are simply issues I discovered while auditing formatting changes.