[ Server ] Send to a client a different websocket server url if provided  (#11428)
## Description

At the moment, clients connect to Alfred to create documents using the REST interface and then connect again with WebSockets to send and receive deltas. 
This PR added an option to provide a URL different from the orderer to connect to the delta stream WebSocket server URL. 

WebSocket server can be specified in the config.json file. If not provided, the WebSocket server URL defaults to Alfred URL.