Fix issues related to Removed Segments in MergeTree (#11678)
Handling removed segments in MergeTree was incorrect, specifically when
to consider them in the case of conflict. This case was complicated by
the fact that clients can have different removed segments depending on
their lifetime. For instance, a new client loading from summary could
have no removed segments, while long running clients could have many and
different removed segments yet to be cleaned up by zamboni.

Previously there was logic added to treat the length of some deleted
segments as undefined. This change was made as any segment of length 0
is considered when resolving a conflict, and this allowed the merge tree
to differentiate removed segments from segments with sequence numbers
greater than the reference sequence number of some incoming operation.
This improved the situation, as it allowed us to exclude those segments
from conflict resolution.

However, the heuristic to determine when to consider a removed segment
for conflict resolution or not was incorrect, and that is what is
improved upon here. Specifically, removed segments who's removed
sequence number in within the collab window, are now consider for
conflict resolution, and those outside the collab window are not. One
key point is we do not consider the removed sequence number itself in
conflict resolution, we only consider the sequences insert sequence
number as segment order has always been determined by insert order which
is an invariant we cannot break.

This works because all clients, regardless of lifetime must agree on the
state of all segments in the collab window, as these segments are not
yet finalized in the sense that they could be above or below the
reference sequence number for other clients, so all clients must be able
to reason over them.

This change also brings in some new unit tests that would previously
fail, and does a bit of unit test refactoring to make them more
consistent.

For now, i've put this new length calculation under and option, as it
will regress some existing assumptions around undo redo, but to fix undo
redo i want to be able to test with this change

fixes https://github.com/microsoft/FluidFramework/issues/9703 
related https://github.com/microsoft/FluidFramework/issues/1009
[AB#45](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/45)
[AB#841](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/841)

Co-authored-by: Abram Sanderson <Abram.sanderson@gmail.com>