Improve Sequence API documentation (#11788)
## Description

Adds and cleans up doc comments on `@fluidframework/sequence` and
`@fluidframework/merge-tree`. Primarily focuses on DDS methods and
documenting semantics of `IntervalCollection`.

This doesn't attempt to document all methods, just provides a start for
more relevant ones in the API surface. Notable areas that could still
use improvement are:

- Sequence event semantics
- IInterval, Interval, and SequenceInterval

This PR is into next because there are some ongoing cleanups of the API
surface in merge-tree and sequence. It also cleans up
waitIntervalCollection and getTextRangeWithPlaceholders, which were some
deprecated APIs that are straightforward to migrate off of but were
never removed in next.

The API reports in many cases suffer from [this API-extractor
bug](https://github.com/microsoft/rushstack/issues/3462) which notes
more things as undocumented than should be noted.

Co-authored-by: Abram Sanderson <absander@microsoft.com>