TreeDDS: verifyChangeRebaser (#11743)
## Description

This PR aims to provide a function to test the changeRebaser class when
provided with a ChangeRebaser object, set of changes, and an equivalence
function. Currently the test cases in "testChangeRebaser.spec.ts"
created are for the "counter" field kind.

## Reviewer Guidance

### Questions
1. A number of inputs (i.e. infinity, NAN, MAX_INTEGER, etc. ) were used
to provide breaking examples for the requirements. However, one of the
requirements (sandwich rebase) does not seem to fail with any of these
inputs.

`(A ↷ B) ↷ [B⁻¹, B] === A ↷ B`

Would there be any example inputs with the counter field kinds which can
break this requirement? Or would it be best to create a rebaser object
which may have "incorrect" rebase logic, but designed to make this
requirement fail (just for testing purposes)?

Co-authored-by: Yann Achard <achardy@microsoft.com>