Random Sequence Generator (#11742)
## Description

This PR aims to create a function to generate random changesets via the
`sequenceEditBuilder`. The following functions were created:
1. `generateRandomUpPaths` - generates random set of UpPaths to
represent a tree
2. `generateRandomChange` - generates a random change operation via the
`sequenceEditBuilder`
3. `changeCombinator` - takes in a random change generation function,
and combines the generated changes to create a more complex changeset
through `rebase`, `compose` and `invert` operations.

## Reviewer Guidance

### Questions
1. The following import (in file `randomSequenceGenerator.ts`) seems
like it will require the `package.json` file of the tree directory to be
updated. Would there be a better location for this file so that we could
avoid updating the `package.json` file?
```
import { makeRandom } from "@fluid-internal/stochastic-test-utils";
```

Co-authored-by: Yann Achard <achardy@microsoft.com>