Tree: cleanup shared-tree test and fix double editing bug (#11850)
## Description

This fixes the bug found in #11838, and cleans up the test to actually
be able to catch the bug. Also fixes a leak (which does not really
matter in that test, but its good practice not to leak things,
especially in tests people use as example usage).

This fix works by having transaction roll back its change after
temporarily applying it during the transaction. The edit then gets
reapplied when the local edit is added.

Also included is a fix that aborted edits would have gotten applied and
submitted.