Lie to ContainerRuntime about connected state when in readonly mode (#11655)
#
[AB#1246](https://dev.azure.com/fluidframework/internal/_workitems/edit/1246)

For more information about how to contribute to this repo, visit [this
page](https://github.com/microsoft/FluidFramework/blob/main/CONTRIBUTING.md).

## Description

`forceReadonly` is rejoining as write and causing new ops to send
`deltaManagerReadonlySubmit` error.
This happens because there are some pending ops that are getting sent at
the same time we are setting `readonly`.

We don't want these ops to be sent yet and want to instead have them
remain as pending. This is done by lying to the `ContainerRuntime`
saying we are `disconnected` whenever we are in `readonly` mode.