Return IDataStore instead of IFluidRouter
In an upcoming change these methods (and others) are going to return an entry point with a handle, but I need to consolidate them on the same return type/interface as ContainerRuntime.createDataStore and ContainerRuntime._createDataStoreWithProps so they all change at the same time.

IDataStore extends IFluidRouter so this shouldn't be breaking.