Shifting message parsing responsibility to container runtime (#11811)
This change prepares code base to shift responsibility for message content parsing from loader to runtime layer.
This helps for two kinds of processes:
- compressed payloads - they are base64 strings now, binary payload latter. No need to wrap with JSON
- chunked ops - wrapping with JSON quite often doubles payload size due to escaping.

It also lets all ops go to runtime, preparing code base for removal of cross-layer reaching. I.e. in future releases we will remove ability of runtime to reach out to DeltaManager "op" handler.
Runtime is in the right position to expose proper information as it sees fit (specifically this applies to summary-related ops).

Validated that runtime can parse ops just fine if Loader stops parsing. This change will be made in next branch for back-compat reasons.

Please note that "op" handler in runtime now has extra arg to indicate if it's a runtime message or not.
For now, for back-compat reasons, it get invoked only for runtime messages, but that's going to change in the future.

Opened the following bugs to track issues discovered in Pending State manager and dirty flag management:
- https://dev.azure.com/fluidframework/internal/_workitems/edit/1832
- https://dev.azure.com/fluidframework/internal/_workitems/edit/1833