Add logging when uploading Summary (#11902)
In order to properly address [bug1871](https://fluidframework.visualstudio.com/internal/_workitems/edit/1871)
we need to understand whether FRS is getting the correct proposal handle
when we upload it. This will exclude the possibility of having AFR
sending the same proposal handle 2x and will allow us to understand why
we are processing a previous handle at all. Once we log the telemetry on
our stress tests, the plan is to remove this logging.