Read data store base summary during creation and update summarizer node. (#11809)
## Description
This change does couple of things:
1. Removes the compat code in remote data store context that reads and converts base summary in "string" format. All the call sites (in dataStores.ts) pass in ISnapshotTree or undefined so its not possible for this to a string anymore.
2. When a data store is created, it updates the summarizer node with its base summary. Here's why this is needed:

When a data store is realized, it reads the base summary it loaded from and updates state from it in the summarizer node. For instance, it reads whether the base summary has DDSes under ".channels" sub-tree and updates this state which is later used when generating a SummaryHandle for this data store.

If application code as part of data store realization, it's possible that data stores are realized out-of-order. What this means is that a data store can be realized after it has been summarized. Consider the following scenario and how this affects the state of the data store (and eventually its children):
- A container with 2 data stores is loaded from a summary. A new summary process starts.
- Data store 1 is summarized. It does not have any changes so it is not realized as part of the summary process. It's ".channels" state is not updated, i.e., it does not know whether in the previous summary, its childern are under ".channels" sub-tree or not. This non-updated state is cached as the work-in-progress state for this summary.
- Data store 2 is summarized and as part of running the application code, it loads data store 1 resulting in its realization. Data store 1's children nodes are created.
- The summary is uploaded to the server and "completeSummary" is called with the summaryHandle. Data store 1's work-in-progress state is stored in a pending queue as pending summary state. It updates the pending state of its children state with its pending state which was not updated with the ".channels" information.
- Summary ack is received and pending summary state becomes latest summary state for data store 1 and its children nodes.
- In the next summary, if any child nodes of data store 1 did not change, they end up sending SummaryHandle with incorrect ".channels" information encoded resulting in the summary being rejected.

The scenario described above is repro'd in the tests added by this PR which fixes this bug as well - https://github.com/microsoft/FluidFramework/pull/11697. Although the bug was fixed, this has the potential to cause other issues because the data store's base summary state is not updated when it its summarized.
A data store's base summary should always be read and its summarizer node's state be updated from it.

## Does this introduce a breaking change?
Althought this changes APIs in the summarizer node, it is confined to the container runtime package. Only container runtime and data store context calls the API that is changed so there are no back-compat concerns here.

[AB#1817](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/1817)