[build-tools] Refactor and better organize the release state machine and handlers (#11834)
Makes the following changes to the `flub release` command, its
underlying state machine, and the classes/interfaces used.

- Handling the state machine states is now done by a dedicated state
handler, and command classes contain a handler property that they call
rather than implementing the StateHandler interface on the command
class. This makes it easier to test command logic vs. state handler
logic.
- State handlers still use inheritance and pass unhandled states to
their parent class for handling.
- Command classes define their machine and the state handler to use.
- When handling a state, a state handler is expected to check the
`testMode` flag and if it is set, immediately return true. This flag is
used in testing to check that all state machine states are handled. Note
that this does not check that the states are handled correctly -- just
that they are handled in some way.
- I've moved most of the logic for handling each state from within the
command's switch statement (the "router") into individual functions that
all share the same signature. This should make it easier to code review
and maintain, because the switch statement will be much smaller. The
switch statement still acts as a router, but each function is now pure
(albeit with a lot of inputs).
- I'm not sure about the organization of the files, particularly the
handler functions and state machines. I've put the state machine stuff
in one place and the handlers in another, but maybe a different
structure would be better.
- I stopped exporting most of the API; we can export it later if needed.
It was handy to see the signatures for everything in one place, though.