Only fail passing tests due to log error (#11915)
## Description

Previously it was possible for a test to fail, and then the after each
to also fail due to logging error. This change only fails a test in the
after each if the test otherwise passed. This will make the logs easier
to read, and the test results easier to understand, as the logs will
only cause the test to fail if it hit no other errors.