[build-tools] Add release report command (#11914)
Adds a `release report` command to the build-cli. This command is used
to collect version information for release groups and packages and
produce a JSON file with data about the release. After a release, it is
useful to generate this report to provide to customers, so they can
update their dependencies to the most recent version.

The command will prompt you to select versions for a package or release
group in the event that multiple versions have recently been released.