MergeTree: Revertibles and Removed Segment Tracking via Local Reference (#11899)
## Description

This change introduces Revertibles to the Merge-Tree package. This
co-locates the core revert logic, which can be used to build undo redo
in the merge tree package. This improves testability, and reduces
concept leakage from the package, as would be necessary to implement
elsewhere.

The primary change here is in how we track and revert removed segments.
Previously it was done via tracking the deleted segments, however this
only worked due to bugs in the handling of removed segments and had
negative impacts on the tree itself in terms of fragmentation. For now,
the previous way of tracking still works to give people time to move to
the new way. Additionally, the new way requires bug fixes currently
under an option (mergeTreeUseNewLengthCalculations), which also breaks
the old way, so staging the release of this will need to be done
carefully.

The primary change in this review is contained in the revertibles file.
For insert and remove the code aligns with the existing undo redo
implementation. For remove it is substantially different. The primary
functions for remove are appendLocalRemoveToRevertibles and
revertLocalRemove. All other changes are in support of the remove
function, or tests for revertibles in general, we were severely lacking.
The most thorough comes from revertibles.farm.spec.ts. There are simple
smoke tests for add, remove, and annotate is both merge tree, and
sequence, along with regression test for issues found with
revertibles.farm.spec.ts.

## Remove Revertible and Local References

As before, the append functions are event based. The primary difference
in the new append for remove is rather than creating a tracking group on
the removed segment itself, it adds a sliding local reference to the
segment. The local reference contains two properties, 1) the serialized
form of the segment, so it can be re-inserted, 2) a tag (referenceSpace)
so the reference can be identified as being created for revert. We also
configure the afterSlide callback for this reference, and in the case
that this reference is detached, we move it to a new kind of segment,
the EndOfTreeSegment. This local reference is then added to a tracking
group for tracking.

As mentioned above, this change brings in a new kind of segment, the
EndOfTreeSegment. This is a special segment that is not bound or known
by the merge tree itself, but it pretends to be a removed segment at the
end of the tree. It is removed so it appears as a undefined/0 length
segment. This segment adds the capability to hold local references that
have been detached from the real merge tree, and give the appearance
that they exist at the end of the tree. In general, local references
only become detached when the tree becomes empty and the
EndOfTreeSegment allows us to gracefully handle that case by giving
those references a place to live. The EndOfTreeSegment is not in itself
a generalized solution to the problem of detached references, as it
requires special and careful understanding of its usage. Specifically,
it is local only and would not lead to eventually consistent results if
serialization of its contained references was attempted, as on
deserialization a naive re-insertion would not put the back into the
EndOfTreeSegment leading to inconsistent results with existing clients.

The last part of the revert remove lifecycle is the re-insertion of the
removed segment by revertLocalRemove. The first phase of the re-insert
is to find the appropriate position. For this we first get the position
of the tracked local reference-- the EndOfTreeSegment give a correct
result here. After we get the position, we need to see if we must shift
it. We only need to shift if the local reference's segment is not
removed, and the local references has been slid to its after tombstones,
as if it is in the after tombstones, then the original segment was after
its current position, and inserts come before by default. We don't need
to shift on removed segments, as they have a length of 0, so after or
before is meaningless, and shifting would put into the wrong positions.

After we get the position, we insert the segment as normal, but we are
not done, as we need to split or shift any adjacent local references
related to revertibles such that we account for the re-insertion. First,
we decide if we will move forward or backward from the new segment to
the local reference it was inserted for. We do this based on the
ordinal, as the ordinal provides are assigned per segment such that they
can be lexicographically sorted. We then walk all segments, and all
references of those segments from the re-inserted segment to the
original local references, and slide any encountered local references
that are part of the revertible reference space to the newly inserted
segment. This preserves the ordering of all the revertible references in
the presence of the re-insert. Generally, this walk will only cover a
small number of segments which are the result of local removes, and only
segments of unacked local removes will contain any local references. In
the absence of removes, the walk will simply split the references on the
local refences we inserted for segment.

Lastly, we have special handling for the EndOfTreeSegment if the
inserted segment was a reference on the EndOfTreeSegment. The
EndOfTreeSegment will not be seen by the above walk, as it is not part
of the tree. In that we manually shift the preceding local references


[AB#45](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/45)

[AB#27](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/27)